Update timeline and about us content in sitetext.yml
The content for the timeline and about us sections in the sitetext.yml file is replaced with a more detailed and elaborate information. This better represents the journey and offerings of Carlos Weeks. The clients section has been removed and irrelevant links in the navigation.yml file have been deleted.